Implement large file streaming for proxy GET
NOTE: This is a backport of the attempt to implement file streaming on the archive-enabled version of this service. I've cut out digest handling and tracking record setup to make it work in the older codebase. It's based on commit: 989c6c165741d0ec20d7b6cf8611187a9e3a9adb.